Add gameController module
This code adds the foundation of the gameController object which contains
most of the 'behind-the-scenes' logic of the game.

It can generate a new game board based on provided size parameters, it
starts and ends the game as well as each individual turn. Finally, it
also handles the player objects, assigning them 'markers' they will play
the game with and adds wins to their win counter after the game is over.