Refactor menu screen display logic
Previous system for switching between different menus relied on setting
a class of 'hidden' with 'display: none' on menus that are no longer in
focus. This caused issues with inline SVG reused throughout the page,
due to the way HTML treats IDs inside of <defs> elements. If an SVG ID
had 'display: none' applied to it anywhere in the page, it could no
longer be accessed from anywhere else.

The new system completely removes unused menu screens from the DOM, thus
solving the issue.