Fix 'avatar selection' display in Firefox
By default Firefox doesn't allow setting a ::before element on radio
inputs, causing the previous version of the page to display incorrectly.

Another issue was the fact that Firefox ignores height and width = 0
declarations on the radio element, instead displaying the radio button
with width and height of 4px. This can be worked around with
visibility:hidden but that might cause accessibility issues for screen
reader users.

The current solution fixes both of those problems.