Fix 'Victory' screen animation
Profile icon was accidentally inheriting a declaration of
margin-right:auto; which affected the positioning of the element without
my knowledge. With a long username, the icon would be displayed aligned
with the left side of the container instead of being centered.

Previous declaration was overwritten and new positioning values put in
place to fix the problem.